+++ fluttercow/src/main/java/org/sralab/fluttercow/Game.java
@@ -201,7 +201,7 @@ public class Game extends EmgImuBaseActivity {
     }
 
     @Override
-    public void onEmgBuffReceived(BluetoothDevice device, int count, int[][] data) {
+    public void onEmgBuffReceived(BluetoothDevice device, int count, double[][] data) {
     }
 
     @Override
