+++ streaming/src/main/java/org/sralab/emgimu/streaming/Streaming.java
@@ -182,7 +182,7 @@ public class Streaming extends EmgImuBaseActivity {
     }
 
     @Override
-    public void onEmgBuffReceived(BluetoothDevice device, int count, int[][] data) {
+    public void onEmgBuffReceived(BluetoothDevice device, int count, double[][] data) {
         if (mAdapter != null) {
             mAdapter.onBuffValueReceived(device);
         }
